+++ TP1/src/parque/PromocionAxB.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class PromocionAxB extends Promocion {
 	
-	public PromocionAxB(List<Atraccion> atracciones, int vigencia, Atraccion atraccionGratuita) {
+	public PromocionAxB(List<Atraccion> atracciones, float vigencia, Atraccion atraccionGratuita) {
 		super(atracciones, vigencia);
 		
 		this.criterioDeDescuento = atraccionGratuita.getCosto();
